tables.php : Added Modal in for each loop so each button has it's own modal with a button that has the record ID attached.
This commit adds functionality to the deleteRecord function and introduces a new function refreshPage.

In the deleteRecord function:

An AJAX request is made to the deleteRecord.php script, sending data including the unique ID, table name, and column name.
Upon success, if the response indicates success, the modal corresponding to the unique ID is hidden, and the success modal is shown.
If an error occurs during the AJAX request, the error is logged, and the error modal is shown with the error message.
The refreshPage function:

This function simply reloads the current page, effectively refreshing it.
These changes enhance the functionality of the application by providing feedback to the user when a record is successfully deleted or if an error occurs during the deletion process. Additionally, the refreshPage function offers a way to reload the page if needed.

Creted deleteRecord.php file that connects and deletes record.

Updated thaeter name in database so TheaterFunctions was updated to have the same name.