Create ShowObj which is used to get show times using MovieID. Created ShowTimeFunctions file. These PHP functions are used to retrieve movie showtimes from a database.
The getMovieTimes function takes two parameters, $movieID and $date, and returns an array of ShowObj objects representing the showtimes of a specific movie on a given date. It prepares a SQL query to select showtimes based on the movie ID and date, executes the query, and fetches the results. It then creates a ShowObj object for each row in the result set and adds it to the $Showtimes array.

The getMovieShowTimeDates function takes a single parameter, $movieID, and returns an array of unique dates on which the movie is being shown. It prepares a SQL query to select showtimes based on the movie ID, executes the query, and fetches the results. It then iterates over the result set, extracting the dates and adding them to the $dates array while ensuring that each date is unique.

Overall, these functions provide convenient interfaces to query and retrieve movie showtime data from the database, allowing for easy integration into PHP applications.

Created SelectTheater page which user will use to purchase tickets.