Added function to Customer Function file, Modified code to forgotPassword.php, and added a new Ajax function to Navbar partial which will be used to start a forgotten password process.
updateToken: This function updates the token and its expiry date for a given email in the Customers table of a database. It generates an expiry date one hour from the current time, prepares and executes an SQL statement to update the token and expiry date, and returns true if the update was successful, otherwise false.

forgotPassword: This function retrieves an email from a form input field, sends it to the server using AJAX to a PHP script (forgotPassword.php), and handles the success and error responses. It logs the server's response and can be further extended to perform actions upon successful insertion.