Updated Customers to Users. All functions inside CustomerFunctions now uses users as they key phrase.
Created a new Object called MoviePrices, which will be used to get ticketPrices using the movieID session variable.

Added a session variable check to see if existing user is an admin. When the user is logged in the type of account will be check and will set a session variable called "IsAdmin"